refactor: renamed matcher to manifest
