+++ .buildkite/pipeline.env.yml
@@ -6,7 +6,6 @@ env:
   MONOFO_DESYNC_STORE: "s3+https://s3-us-west-2.amazonaws.com/vital-buildkite-cache-us-west-2/desync/store"
   MONOFO_DESYNC_CACHE: "/tmp/desync/monofo"
   NODE_ENV: development
-  MONOFO_HOOK_DEBUG: 1
 
 monorepo:
   matches: true
